backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
openai==1.3.0
httpx<0.28,>=0.25.0
httpcore<1.0,>=0.18.0
python-dotenv==1.0.0
pydantic==2.5.0
numpy>=1.24.3,<2.0.0
sentence-transformers>=2.2.2
scikit-learn>=1.3.2
rake-nltk==1.0.6
keybert==0.8.1
bertopic==0.15.0
neo4j==5.14.0
python-multipart==0.0.6
aiofiles==23.2.1
torch>=2.0.0
transformers>=4.30.0

